requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
django-compressor
django-widget-tweaks
Pillow